Setup and cleanup
Wanted to continue the series and implement some of the effects and new classes.
Unfortunately I just found out whatever changes I made the Player was now a (Demi)Lich going for the infinite stones. Even while dead he was never freed by the area.
To figure out what went wrong I created some helper functions and decides to give every object a unique name to make the deugging process easier.
At the end it was just a indexing  error that occured when changing the iterating in the Area from start->end to end->start. As `range(n, 0, -1)` results in [n, ... 2, 1] and not [n, ... 1, 0] as I was expecting.

* .git_commit_msg

   Changed commit workflow to `git commit -F .git_commit_msg`. So the changes will be tracked now while doing it in a more detailed way here. Therefore the file was added to the `.gitignore` as well.

* Scenes/Stage.gd

   DRY - `InputMap` controls by getting the common parts into `add_inputs()` but only works for `InputEventKey` (`InputEventJoypadButton` etc need still their own function , but are not used in GdPath)

* Ship-Beta.png, Ship-Beta.png.import, icon.svg, icon.svg.import

   Removed the default Godot-Logo and added a custom dummy one just for learning how to

* Player.gd

   - Destroy if the Player is out of bounds (Attention: that does not consider Camera.shake()s amplitude)

* Area.gd

   - added `list_gameobject_types()` to return a dictionary with the classes and the number of objects currently in the area
   - fixed a bug with `_process(_delta)` as the first object added was never checked

* Misc

   - every GameObject has now a name it will be its class_name + # provided by `G.get_id()` (TODO can this be automated?)